Temporary fix for valid mask (#54)
* Temporary fix for valid mask

* unit test bug fix

+++ openoptflow/data/dataset/base_dataset.py
@@ -94,9 +94,9 @@ class BaseDataset(data.Dataset):
         if self.augmentor is not None:
             img1, img2, flow = self.augmentor(img1, img2, flow)
 
-        img1 = torch.from_numpy(img1).permute(2, 0, 1)
-        img2 = torch.from_numpy(img2).permute(2, 0, 1)
-        flow = torch.from_numpy(flow).permute(2, 0, 1)
+        img1 = torch.from_numpy(img1).permute(2, 0, 1).float()
+        img2 = torch.from_numpy(img2).permute(2, 0, 1).float()
+        flow = torch.from_numpy(flow).permute(2, 0, 1).float()
 
         return (img1, img2), flow
